♻️ refactor:  EnvをUtilsに移動
Issue #278
