+++ lib/infrastructure/github/http_client.dart
@@ -8,9 +8,9 @@ import 'dart:io';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:http/http.dart' as http;
 
-import '../../config/env.dart';
-import '../../config/env_define.dart';
 import '../../domain/exceptions.dart';
+import '../../utils/env/env.dart';
+import '../../utils/env/env_define.dart';
 import '../../utils/logger.dart';
 
 /// GitHubアクセストークンプロバイダー
