+++ test/presentation/pages/repo/components/search_repos_query_test.dart
@@ -5,9 +5,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:github_search/config/env.dart';
 import 'package:github_search/config/router.dart';
 import 'package:github_search/presentation/pages/repo/components/search_repos_query.dart';
+import 'package:github_search/utils/env/env.dart';
 import 'package:go_router/go_router.dart';
 
 import '../../../../test_utils/test_agent.dart';
